Update README.md
Challenge/Opportunity: Provided a detailed account of the challenges faced by Toys "R" Us, framing the problem within the context of the organization's slow adoption of e-commerce and outdated product marketing strategies. Included a footnote that summarised Toys "R" Us's brief description and its mission/purpose. Rationale: Our solution is explained clearly, showing a deep understanding of how data analytics can address the identified needs. The rationale for targeting both physical and online sales, informed by customer demographics and purchasing behavior, is logical and clearly ties into the specific challenges outlined. Cost-benefit Analysis: We went one step ahead and discussed the cost + benefit instead of only cost analysis. The cost-benefit analysis comprehensively covered a detailed breakdown of the initial investment and ongoing costs. We included quantitative metrics such as NPV, ROI, and IRR, which substantiate the project's financial viability. This section aligns with the criteria by identifying costs, required resources, and expected returns. Went above and beyond. Assessment: Described a set of KPIs and monitoring measures to evaluate the project's performance immediately and over time. This forward-looking approach demonstrates a plan for ongoing assessment beyond the initial implementation. Conclusion: Our conclusion effectively reiterates the proposal's urgency and the significant financial benefits, presenting a compelling call to action for the project's approval. It summarizes the key points and reinforces the strategic move to secure a competitive advantage for the company in the short-term and solidify the position for the long-term.